requirements.txt
langchain_groq
langchain_community 
dotenv
langchain_core
streamlit
fastapi
uvicorn 
langserve
sse_starlette
pydantic
sentence-transformers
faiss-cpu
wikipedia
arxiv
langchain-ollama
phidata
streamlit
google-generativeai
duckduckgo-search
youtube-transcript-api
py2pdf
whisper